Zero allocation ZipArchive creation with deferred RangeReader
The allocation of the ZipArchive's comment strikes me as antithetical
with the rest of the crate that tries to surface abstractions that
allows users to avoid heap allocations and in general, follow don't pay
for what you don't use philosophy.

This is a breaking change, `ZipArchive::comment()` now returns the newly
introduced `RangeReader<&R>` instead of `ZipStr`, which exposes a `Read`
interface between offsets for users to read the zip comment into a user
provided buffer. A `RangeReader` is also useful abstraction for
incrementally reading a potentially large prelude in a zip file.

This commit enforces consistency in the slice archive and reader archive
that both validate enough data for the comment is present prior to
returning the created archive. There is room in the future for making
this configurable

With these changes, there are zero allocations along the happy path for
ZipArchive creation.