Expose the end offset of an archive to extract trailing data
This commit adds `ZipSliceArchive::end_offset` and
`ZipArchive::end_offset`, which returns the offset where the ZIP archive
ends.

This returns the position immediately after the last byte of the ZIP
archive, including the End of Central Directory record and any comment.
This is useful for extracting trailing data.

The calculation does not rely on any self reported values from the
archive, so in the case that a comment is truncated in the stream, the
end offset will represent what the underlying reader actually contains.

This can be used in conjunction with base offset to determine the exact
byte range (and thus size) of the ZIP archive within the larger file.

Originally I was going to prefer exposing a "size", but in the context
of zip, size can be ambiguous -- does it represent the size if the
archive is fully inflated? Or what about the deflated size? Or what
about the size of archive and all of internal structures?