Expose central directory file extensible data fields
Until now, the extensible extra data fields have been kept internal and
not exposed. Internally they are used for applying corrected zip64
information and extracting timestamp information. It is high time for
this data to also be exposed externally.

The solution is to add a method `ZipFileHeaderRecord::extra_fields`
which is an iterator over the extra field data.

```rust
let archive = ZipArchive::from_slice(data)?;
for entry_result in archive.entries() {
    let entry = entry_result?;
    let mut extra_fields = entry.extra_fields();
    for (field_id, field_data) in extra_fields.by_ref() {
        match field_id {
            ExtraFieldId::JAVA_JAR => {
                println!("Handle jar CAFE field with {} bytes", field_data.len());
            }
            _ => { }
        }
    }

    // If desired, check for truncated data
    if !extra_fields.remaining_bytes().is_empty() {
        println!("Warning: Some extra field data was truncated");
    }
}
```