Expose end of central directory offset on zip archives
[PR 76][pr] added the end of central directory signature offset in
errors when locating a zip to facilitate false eocd recovery. This
commit takes the thought one step farther by exposing the eocd offset on
the zip archives themselves, as it is fairly easy for the end of central
directory to "pass validation" (so to speak) and get archive created
but turns out to be nonsensical.

[pr]: https://github.com/nickbabcock/rawzip/pull/76