Add EOCD validation on central directory offset
If the self-reported central directory start offset is past the start of
the end of the central directory, then an error should be raised before
a zip archive is even created. This fixes a panic in ZipSliceArchive.

This commit also uses the central directory size declared in the zip64
eocd when available. This should only have an impact when the size of
the central directory is larger than 4GB.

Along for the ride are some improvements to naming and internal
structure for the end of the central directory.