Avoid MutexReader for files on Windows
Windows also supports a pread()-esque function, with the exception that
the file position does get changed. Multiple threads can call it at the
same time, but the file position afterwards is unspecified. This should
not matter for rawzip since it takes ownership of the File instance. The
only scenario where this could be observed externally is if the internal
HANDLE object was duplicated with DuplicateHandle() (File::try_clone())
before the File was passed to ZipArchive. However, Rust's try_clone()
documentation mentions this explicitly and it's unlikely that someone
would pass a File instance to a library and expect the seek position to
be unchanged.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>