Remove ZipVerifier ties to a ZipArchive
This decouples ZipVerifier from a ZipArchive, removing an unncessary
lifetime in downstream crates. Very similar to changes made to ZipReader
in #85

This also sets up ZipVerifier to have a future public API where anyone
can pipe in a decompressor + ReaderAt and be able to verify the size and
crc of an entry. I'm not anticipating that being a strong use case but
so I'm deferring it to a later date.